preview for images and files in chat
